requirements.txt
# Packages must be installed in prod and in dev, so they are listed here.
# The requirements-dev.txt file includes this file and adds additional packages that are only needed for development and testing.
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
Django==5.2.11
djangorestframework==3.16.*
djangorestframework-simplejwt==5.4.*
django-cors-headers==4.7.*
django-celery-beat==2.8.1
django-crispy-forms==2.5
django-mathfilters==1.0.0
django-csp==3.8
django-ipware==7.0.1
django-ratelimit==4.1.0
django-stubs-ext==5.2.9
django-timezone-field==7.2.1
gunicorn==25.1.0
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0